refactor(hickey): extract shared extractOffsets helper
`offsetOf` in `extractQuote.ts` and `lineRangeForSelection` in
`useTextSelection.ts` were walking text nodes with identical
TreeWalker-and-accumulator scaffolding. Extract one helper
`extractOffsets(root, range)` in `@kolu/artifact-sdk/core` and route
both call sites through it.

Behavior preserved: cross-root ranges still fall back to the old
"slice from text end" semantics in extractQuote (fact-check follow-up
will tighten that to empty prefix/suffix in a separate change).
