+++ packages/artifact-sdk/src/core/extractQuote.ts
@@ -6,6 +6,7 @@
  *  Used by both the parent-side `useTextSelection` and the in-iframe SDK
  *  (bundled by esbuild at server startup, served from the same source). */
 
+import { extractOffsets } from "./extractOffsets";
 import type { Locator } from "../types";
 
 const CONTEXT_WINDOW = 32;
@@ -20,34 +21,6 @@ function rootTextContent(doc: Document | ShadowRoot): string {
   return doc.textContent ?? "";
 }
 
-/** Return the character offset of `range.startContainer` (at `startOffset`)
- *  within the concatenated `rootTextContent(doc)`. Walks text nodes in
- *  document order and accumulates lengths until the start node is hit. */
-function offsetOf(
-  range: Range,
-  doc: Document | ShadowRoot,
-  which: "start" | "end",
-): number {
-  const target = which === "start" ? range.startContainer : range.endContainer;
-  const targetOffset = which === "start" ? range.startOffset : range.endOffset;
-  const root = doc instanceof Document ? (doc.body ?? doc) : doc;
-  const ownerDoc = doc instanceof Document ? doc : doc.ownerDocument;
-  if (!ownerDoc) return 0;
-  const walker = ownerDoc.createTreeWalker(root as Node, NodeFilter.SHOW_TEXT);
-  let acc = 0;
-  let node: Node | null = walker.nextNode();
-  while (node) {
-    if (node === target) return acc + targetOffset;
-    acc += node.textContent?.length ?? 0;
-    node = walker.nextNode();
-  }
-  // Range boundary not inside this root (e.g. cross-root selection) —
-  // fall back to the accumulated total so the prefix is the full text
-  // before. The locator still carries the quote, which is the durable
-  // anchor; prefix/suffix is purely a disambiguator.
-  return acc;
-}
-
 /** Build a Locator from a non-collapsed Range. Caller guarantees non-empty. */
 export function extractQuote(
   range: Range,
@@ -55,9 +28,18 @@ export function extractQuote(
 ): Locator {
   const quote = range.toString();
   const text = rootTextContent(doc);
-  const start = offsetOf(range, doc, "start");
-  const end = offsetOf(range, doc, "end");
-  const prefix = text.slice(Math.max(0, start - CONTEXT_WINDOW), start);
-  const suffix = text.slice(end, end + CONTEXT_WINDOW);
+  // Cross-root ranges: fall back to the full text length as the offset
+  // so the prefix slice ends at the document end. The locator still
+  // carries the quote, which is the durable anchor; prefix/suffix is
+  // purely a disambiguator. (Fact-check follow-up may tighten this.)
+  const offsets = extractOffsets(doc, range) ?? {
+    start: text.length,
+    end: text.length,
+  };
+  const prefix = text.slice(
+    Math.max(0, offsets.start - CONTEXT_WINDOW),
+    offsets.start,
+  );
+  const suffix = text.slice(offsets.end, offsets.end + CONTEXT_WINDOW);
   return { quote, prefix, suffix };
 }
